Adds db support for slider tracks

+++ HathorData/SliderTrack.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[SliderTrack]
 (
 	[SliderID] VARCHAR(20) NOT NULL PRIMARY KEY, 
-    [Duration] TINYINT NOT NULL, 
+    [Duration] INT NOT NULL, 
     [FullTitle] VARCHAR(200) NOT NULL, 
     [Url] VARCHAR(250) NOT NULL, 
     [ExtraInformation] BIT NULL DEFAULT 0
